The rtbkit integration test can now shutdown properly.
It was incapable of shutting down properly because of an issue with
the router's monitor whereby it gets stuck in zmq when trying query
the state of the global monitor which its not connected to yet. This
problem is actually present in many other component and we're still
looking for a permanent to it.

For this test though, it prevented the temporary redis server from
being killed at the end of the test which is bad. So making sure the
monitor is up before we initialize any of the other components that
use that monitor fixes the problem here.