Sanitized the BiddingAgent's init() and start() behaviour and the agentName interface.
The old version didn't follow our usualy conventions and did
everything in the start() function. Unfortunately it never did call
the MessageLoop's init() function which could have caused all sorts of
shutdown problems.

As for the agentName, it's now generated by the BiddingAgent based on
the serviceName with an option to overwrite for tests. This cleans up
the arguments to the start() function which, again, were not very
standard.