BiddingAgent will now send its config to the AgentConfigurationService when the connection is established.
The old behaviour would only send the config when the user provided it
through the doConfig function. If the configuration service is not
available then the config would be dropped and would not be sent when
the configuration service is finally available.

New behaviour will always send the currently set configuration when we
manage to connect to the AgentConfigurationService which ensure that
we always have a valid configuration available in the current
AgentConfigurationService.