Sending on an unconnected zmq socket will now drop the message.
The rational is that if the message is important then an extra
acknowledgement mechanism should be added on top to ensure that
message was properly received. This also fixes several issues where a
user would call connect but the service is not available.

Note that yes this is starting to look like we're re-implementing tcp
but it's necessary to handle the complexity of the zookeeper discovery
mechanism.