ZmqMultipleNamedClientBusProxy should no longer call the connection handler when the service is unavailable.
Turns out that this is pretty complicated to get right because we
can't call the handler while the lock is held which mean that the
naive forwarding implementation won't work. Instead we need a custom
functor to act as the callback which handles a lock-free state machine
whose behaviour changes depending on whether we're in the locked
section or not.

Note that we should eventually try to simply remove the lock by
funnelling all events (zmq and connect calls) through message loop.