Augmentors will now pipe their responses back into the main message loop thread.
This should avoid race conditions when sending the response on the zmq
socket from a different thread then the one doing the sending. While
we do have locks in the zmq layers on send, there still seems to be
some lingering race conditions within zmq (weird aborts on
receive). Hopefully this change should allow us to remove the locks
from the zmq layer and fix the weird race conditions.