Service names are no longer de-duped and should be uniquified from the command line.
The idea is that RTBKit shouldn't mandate the scheme used for the
service name and instead leave it up to the user to specify unique
service names in the command line.

A strong reason for this change is that when deployed on VMs where the
hostname or even the location could change on a VM restart, the
logical name for that service shouldn't change because it's the same
service.