The master banker will now connect to every monitor present in zookeeper regardless of location.
The setup for multi-dc calls for 1 monitor per data-center and 1
master globally. In order for all data-centers to stop bidding when
the master banker dies, the master banker needs to register itself
with the monitors of each data-center.

While this affect the MonitorProviderClient, by default the connection
scope is localized the the current location which means that the
original behaviour is preserved.