The startdelay field in OpenRTB will now default to a non-conflicting unspecified value.
Note that startdelay is a TaggedInt and not a VideoStartDelay which is
an enum. The enum itself is only use as a reference for a several
magic values that this field can take. This also meant that the
default value for TaggedInt was incorrect and is now set to be the now
disambiguated value of VideoStartDelay::UNSPECIFIED.

Github issue: #64