Adds support for adgroup_id to adx exchange connector.
An agent can now set in the adx provider configuration the adGroupId
parameter, it is of type int64_t and optional. If set on the config it
will always be set in the bid response if the matching bid request
had it declared for the given adslot match data, otherwise (there is
no match) the response will be filtered and a 'No Bid' will be
sent as a response.