ADX connector now handles restricted categories.
Restricted categories need to be declared in the creative provider
configuration. The attribute is 'restrictedCategory' and it is a string
containing numeric ids separated by white spaces. I.e.:
restrictedCategory = "33 58"
If no restricted category applies then it should be set to 0 I.e. :
restrictedCategory = "0"

The doc says that if you are bidding with ads in restricted categories
you MUST declare them in BidResponse.Ad.restricted_category. This
way if a creative declares a restrictedCategory X and the request does
not allow it, it will get filtered and a 'No Bid' will be sent back.
Otherwise if the request declares X as allowed, it will be specified
on the response.