Added MessageLoop blocking mode (RTBKIT-567)
MessageLoop class's main event loop refactored to use the Epoller's blocking functionality when needsPoll is false.  Other changes were made to allow the loop to exit immediately on shutdown.  There is still an unexplained error in the Epoller::processOne() method not being equivalent to the hard polling version; for the moment that is disabled while investigation happens.  RTBkit testsuite passes.