Reduce number of context switches in agents by 10x (RTBKIT-528)
This changeset adds a test program for context switches (that prints how many an idle agent will do), and changes the default bidding agent maxAddedLatency to 5ms from 0.5ms.  It reduces the number of context switches an idle agent makes to 200 from 2000 per second, at the expense of up to 4.5ms of extra bidding latency.