Add ability to turn off log categories by default (PLAT-564)
Adds an extra boolean parameter to the log category that allows it to be turned off by default, to support the 'recompile with verbose logging' use-case.  This could be combined at a later stage with the ability to modify logging via environment variables or a SHM segment and a signal.