Post auction service can now be sharded for scaling.
Sharding is done at the router and ad server level where the correct
shard is chosen using the hash of the auction id modulo the number of
shards provided via the postAuctionShards parameter in the
bootstrap.json.

The post auction loop must also declare itself as a given shard using
the shard command line argument.