Handle RestConnectionHandler asynchronous disconnects
When a RestConnectionHandler processes a response asynchronously, it takes a reference to the
connection and then returns.  Sometime later it will asynchronously write to the connection and
then close it.  As soon as it returns, it loses the lock on the connection, and it is possible
for the other end to break the connection and for it to be shutdown.  In that case, the
connection is dangling and we were getting a segfault.

We now handle that by trapping the disconnect and setting a mutex-protected zombie flag.  All
operations that require the connection also take the mutex to avoid a race condition.  If
a response to the connection is attempted, it will notice that the connection is a zombie and
exit immediately.  This doesn't solve the problem of releasing the connection, but does at
least avoid the segfault.