PLAT-790: Runner: fixed scalability of memory use
This change implement the code from Runner::Task::runWrapper in a separate
program, enabling the new "runWrapper" to call execv right away. This
enables the wrapper process to release any memory pages inherited from its
parent so that old versions of the modified pages do not need to be kept.