HttpBidderInterface: make sure to tag all the impressions.
For BidRequests with multiple impressions, the HttpBidderInterface might
not tag impressions that have no potential bidders. If used with the
RTBKitExchangeConnector and the CreativeIdsExchangeFilter, this might be
bad because the filter expects that every impression contains at least
an "external-ids" or "creative-ids" extension field. For impressions
that didn't satisfy filters, we now tag them with null values